+++ src/test/clojure/com/appsflyer/donkey/routes.clj
@@ -5,13 +5,14 @@
 ; you may not use this file except in compliance with the License.
 ; You may obtain a copy of the License at
 ;
-;     http://www.apache.org/licenses/LICENSE-2.0
+;      http://www.apache.org/licenses/LICENSE-2.0
 ;
 ; Unless required by applicable law or agreed to in writing, software
 ; distributed under the License is distributed on an "AS IS" BASIS,
 ; WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 ; See the License for the specific language governing permissions and
 ; limitations under the License.
+;
 ;
 
 (ns com.appsflyer.donkey.routes
@@ -149,6 +150,11 @@
    :consumes     ["multipart/form-data" "application/x-www-form-urlencoded" "application/octet-stream"]
    :handler-mode :blocking
    :handler      (fn [_req & _args] {})})
+
+(def internal-server-error
+  {:path         "/internal-server-error"
+   :handler-mode :blocking
+   :handler      (fn [_req & _args] (throw (ex-info "Route throws an exception" {})))})
 
 (def blocking-middleware-handlers
   {:path         "/route/middleware/blocking"